Move SC theme to a Gatsby plugin
Each page needs to be wrapped with a styled-components theme provider in
order to share colors. Technically I could import it directly, but that
would preclude some dynamic theme features, or at least those compatible
with IE 11 (not CSS vars).
