+++ .eslintrc.yml
@@ -13,3 +13,6 @@ plugins:
 extends:
 - eslint:recommended
 - plugin:react/recommended
+
+rules:
+  react/prop-types: off
