+++ gatsby-browser.js
@@ -1,0 +1,1 @@
+export { default as wrapRootElement } from './src/theme';
